Implemention of a read-only file system for OSv and modifications to the build system and VFS to allow using the new file system as the root file system.
There are a few reasons why this is an important addition to OSv:
    1. A read-only root file system allows the user to force the hypervisor
       to not allow writing to disk. This means that you can boot multiple
       intances off of a single image file. A use case for this feature is
       something like Redis, allowing you to only have to maintain
       a single image file.

    2. An immutable image will always have the same hash signature. This will
       allow the user of the image to verify that the image file has not been
       tampered with, even after the image has been run.

    3. From Nadav Har'El: "it allows building an OSv image *without* running
       it, which solves all sorts of nasty corner cases in OSv's current
       scripts/build, when it is not desired to *run* the image during build.
       For example, cross-compilation, building on a VM without efficient
       nested virtualization, etc."

The file system itself is titled MFS (originally standing for My File System,
but the name is not very important). It has a limited set of features, but
the reduced feature set allows the implementation and the layout of the blocks
to be very simple. Currently MFS supports regular files/directories and
symbolic links. File and directory names are limited to 63 characters and link
paths are limited to 511 characters.

The layout of the file system is as follows:
    Block 0:       The superblock. This stores identifing information about the
                   intstance of MFS as well as the Block ID for Inode #1.

    Block 1 - N:   The next set of blocks store files, directory entires, and
                   links. *Everything* in MFS is stored sequentially, meaning
                   that there will be no fragmentation. Also, files can be
                   referenced by knowing their starting block id as well as
                   their length.

    Block N - End: The last set of blocks stores the inodes, in sequential
                   order. Due to being ordered, it is possible to know the
                   location of any inode given only its number.

Based on conversations with Nadav, I went back and forth about implementing a
caching mechanism in the MFS driver. After some testing, I opted to not include
a caching mechanism for the following reasons:

    1. The block I/O system (fs/vfs/vfs_bio.h) already has a cache built in
       when using the bread function.
    2. Through testing with the redis-memonly image, I found that over 96%
       (1132/1162) of file reads are first time reads of a block, meaning that
       it could not have been cached (without a much more complex pre-caching
       system). Of blocks that were read multiple times, the block i/o system
       had an 80% cache hit rate (31/39). This seems to be sufficient for the
       time being. Tests were run with MFS as the root file system and multiple
       runs of redis-benchmark against a single running image.

The build system and OSv modifications were extremly minimal. An option was
added to scripts/build titled rootfs that allows specifying mfs or zfs to be
used (with zfs as the default). Once OSV is running, it will attempt to mount
the root file system as MFS every time. Upon failure (in the case of ZFS being
the root file system), it will fall back to trying to mount the file system as
ZFS. Currently, I do not believe it is possible to have both ZFS and MFS
running in the same image (zpool is not included in MFS images) but this should
be implemented in the future.